+++ puma/tests/hlplots/test_aux_results.py
@@ -7,12 +7,10 @@ import unittest
 from pathlib import Path
 
 import numpy as np
-from ftag import get_mock_file
 
 from puma.hlplots import AuxResults
 from puma.hlplots.tagger import Tagger
-from puma.utils import logger, set_log_level
-from puma.utils.vertexing import calculate_vertex_metrics
+from puma.utils import get_dummy_tagger_aux, logger, set_log_level
 
 set_log_level(logger, "DEBUG")
 
@@ -21,49 +19,45 @@ class AuxResultsTestCase(unittest.TestCase):
     """Test class for the AuxResults class."""
 
     def test_add_taggers_from_file(self):
-        """Test for Results.add_taggers_from_file function."""
-        fname = get_mock_file()[0]
-        results = AuxResults(signal="bjets", sample="test")
-        taggers = [Tagger("MockTagger")]
+        """Test for AuxResults.add_taggers_from_file function."""
+        np.random.default_rng(seed=16)
+        fname = get_dummy_tagger_aux()[0]
+        results = AuxResults(sample="test")
+        taggers = [Tagger("GN2")]
         results.add_taggers_from_file(
             taggers,
             fname,
-            vtx_label_var="numberOfPixelHits",
-            vtx_reco_var="numberOfSCTHits",
         )
         self.assertEqual(list(results.taggers.values()), taggers)
 
     def test_add_taggers_with_cuts(self):
-        fname = get_mock_file()[0]
+        np.random.default_rng(seed=16)
+        fname = get_dummy_tagger_aux()[0]
         cuts = [("eta", ">", 0)]
         tagger_cuts = [("pt", ">", 20)]
-        results = AuxResults(signal="bjets", sample="test")
-        taggers = [Tagger("MockTagger", cuts=tagger_cuts)]
+        results = AuxResults(sample="test")
+        taggers = [Tagger("GN2", cuts=tagger_cuts)]
         results.add_taggers_from_file(
             taggers,
             fname,
-            vtx_label_var="numberOfPixelHits",
-            vtx_reco_var="numberOfSCTHits",
             cuts=cuts,
         )
         self.assertEqual(list(results.taggers.values()), taggers)
 
     def test_add_taggers_with_cuts_override_perf_vars(self):
         rng = np.random.default_rng(seed=16)
-        fname = get_mock_file()[0]
+        fname = get_dummy_tagger_aux()[0]
         cuts = [("eta", ">", 0)]
         tagger_cuts = [("pt", ">", 20)]
-        results = AuxResults(signal="bjets", sample="test")
-        taggers = [Tagger("MockTagger", cuts=tagger_cuts)]
+        results = AuxResults(sample="test")
+        taggers = [Tagger("GN2", cuts=tagger_cuts)]
         results.add_taggers_from_file(
             taggers,
             fname,
-            vtx_label_var="numberOfPixelHits",
-            vtx_reco_var="numberOfSCTHits",
             cuts=cuts,
             perf_vars={
-                "pt": rng.exponential(100, size=1000),
-                "eta": rng.normal(0, 1, size=1000),
+                "pt": rng.exponential(100, size=9999),
+                "eta": rng.normal(0, 1, size=9999),
             },
         )
         self.assertEqual(list(results.taggers.values()), taggers)
@@ -74,20 +68,27 @@ class AuxResultsPlotsTestCase(unittest.TestCase):
 
     def setUp(self) -> None:
         """Set up for unit tests."""
-        f = get_mock_file()[1]
-        dummy_tagger_1 = Tagger("MockTagger")
-        dummy_tagger_1.labels = np.array(
+        f = get_dummy_tagger_aux()[1]
+        dummy_tagger = Tagger("GN2")
+        dummy_tagger.labels = np.array(
             f["jets"]["HadronConeExclTruthLabelID"],
             dtype=[("HadronConeExclTruthLabelID", "i4")],
         )
-        dummy_tagger_1.aux_metrics = calculate_vertex_metrics(
-            f["tracks"]["numberOfPixelHits"],
-            f["tracks"]["numberOfSCTHits"],
-        )
-        dummy_tagger_1.perf_vars = {"pt": f["jets"]["pt"]}
-        dummy_tagger_1.scores = f["jets"]
-        dummy_tagger_1.label = "dummy tagger"
-        self.dummy_tagger_1 = dummy_tagger_1
+        dummy_tagger.aux_scores = {
+            "vertexing": f["tracks"]["GN2_VertexIndex"],
+            "track_origin": f["tracks"]["GN2_TrackOrigin"],
+        }
+        dummy_tagger.aux_labels = {
+            "vertexing": f["tracks"]["ftagTruthVertexIndex"],
+            "track_origin": f["tracks"]["ftagTruthOriginLabel"],
+        }
+        dummy_tagger.perf_vars = {"pt": f["jets"]["pt"]}
+        dummy_tagger.label = "dummy tagger"
+        dummy_tagger_no_aux = Tagger("GN2_NoAux", aux_tasks=[])
+        dummy_tagger_no_aux.perf_vars = {"pt": f["jets"]["pt"]}
+        dummy_tagger_no_aux.label = "dummy tagger no aux"
+        self.dummy_tagger = dummy_tagger
+        self.dummy_tagger_no_aux = dummy_tagger_no_aux
 
     def assertIsFile(self, path: str):
         """Check for file to exist.
@@ -106,38 +107,37 @@ class AuxResultsPlotsTestCase(unittest.TestCase):
         if not Path(path).resolve().is_file():
             raise AssertionError(f"File does not exist: {path}")
 
-    def test_plot_var_vtx_eff(self):
-        """Test that png file is being created."""
-        self.dummy_tagger_1.reference = True
+    def test_plot_var_vtx_perf_alljets(self):
+        """Test that png files are being created for tagger with aux tasks."""
+        self.dummy_tagger.reference = True
         with tempfile.TemporaryDirectory() as tmp_file:
-            auxresults = AuxResults(signal="bjets", sample="test", output_dir=tmp_file)
-            auxresults.add(self.dummy_tagger_1)
-            auxresults.plot_var_vtx_eff()
-            self.assertIsFile(auxresults.get_filename("vtx_eff_vs_pt"))
+            auxresults = AuxResults(sample="test", output_dir=tmp_file)
+            auxresults.add(self.dummy_tagger)
+            auxresults.plot_var_vtx_perf()
+            self.assertIsFile(auxresults.get_filename("alljets_vtx_eff_vs_pt"))
+            self.assertIsFile(auxresults.get_filename("alljets_vtx_purity_vs_pt"))
+            self.assertIsFile(auxresults.get_filename("alljets_vtx_nreco_vs_pt"))
+            self.assertIsFile(auxresults.get_filename("alljets_vtx_trk_eff_vs_pt"))
+            self.assertIsFile(auxresults.get_filename("alljets_vtx_trk_purity_vs_pt"))
 
-    def test_plot_var_vtx_fr(self):
-        """Test that png file is being created."""
-        self.dummy_tagger_1.reference = True
+    def test_plot_var_vtx_perf_bjets(self):
+        """Test that png files are being created for tagger with aux tasks."""
+        self.dummy_tagger.reference = True
         with tempfile.TemporaryDirectory() as tmp_file:
-            auxresults = AuxResults(signal="bjets", sample="test", output_dir=tmp_file)
-            auxresults.add(self.dummy_tagger_1)
-            auxresults.plot_var_vtx_fr()
-            self.assertIsFile(auxresults.get_filename("vtx_fr_vs_pt"))
+            auxresults = AuxResults(sample="test", output_dir=tmp_file)
+            auxresults.add(self.dummy_tagger)
+            auxresults.plot_var_vtx_perf(flavour="bjets")
+            self.assertIsFile(auxresults.get_filename("bjets_vtx_eff_vs_pt"))
+            self.assertIsFile(auxresults.get_filename("bjets_vtx_purity_vs_pt"))
+            self.assertIsFile(auxresults.get_filename("bjets_vtx_nreco_vs_pt"))
+            self.assertIsFile(auxresults.get_filename("bjets_vtx_trk_eff_vs_pt"))
+            self.assertIsFile(auxresults.get_filename("bjets_vtx_trk_purity_vs_pt"))
 
-    def test_plot_var_vtx_trk_eff(self):
-        """Test that png file is being created."""
-        self.dummy_tagger_1.reference = True
+    def test_plot_var_vtx_perf_empty(self):
+        """Test vertexing performance function with empty data."""
+        self.dummy_tagger_no_aux.reference = True
         with tempfile.TemporaryDirectory() as tmp_file:
-            auxresults = AuxResults(signal="bjets", sample="test", output_dir=tmp_file)
-            auxresults.add(self.dummy_tagger_1)
-            auxresults.plot_var_vtx_trk_eff()
-            self.assertIsFile(auxresults.get_filename("vtx_trk_eff_vs_pt"))
-
-    def test_plot_var_vtx_trk_fr(self):
-        """Test that png file is being created."""
-        self.dummy_tagger_1.reference = True
-        with tempfile.TemporaryDirectory() as tmp_file:
-            auxresults = AuxResults(signal="bjets", sample="test", output_dir=tmp_file)
-            auxresults.add(self.dummy_tagger_1)
-            auxresults.plot_var_vtx_trk_fr()
-            self.assertIsFile(auxresults.get_filename("vtx_trk_fr_vs_pt"))
+            auxresults = AuxResults(sample="test", output_dir=tmp_file)
+            auxresults.add(self.dummy_tagger_no_aux)
+            with self.assertRaises(ValueError):
+                auxresults.plot_var_vtx_perf()